TorrentService: mess with result type definitions
i want to store the Result type within the class for nice usage,
but I can't define it for the first time inside it as then I can't refer
to it in the inherited type. i hide away the original type alias in an
anonymous namespace so as not to clutter the main namespace, and alias
that to the inner Result type.

weh